// src/app/maps/interfaces/directions.ts
// Generated by https://quicktype.io

export interface DireccionesResponse { 
    routes:    Route[];
    waypoints: Waypoint[];
    code:      string;
    uuid:      string;
}

export interface Route {
    country_crossed:  boolean;
    weight_typical:   number;
    duration_typical: number;
    weight_name:      string;
    weight:           number;
    duration:         number;
    distance:         number;
    legs:             Leg[];
    geometry:         Geometry;
}

export interface Geometry {
    coordinates: Array<number[]>;
    type:        Type;
}

export enum Type {
    LineString = "LineString",
}

export interface Leg {
    via_waypoints:    any[];
    admins:           Admin[];
    annotation:       Annotation;
    weight_typical:   number;
    duration_typical: number;
    weight:           number;
    duration:         number;
    steps:            Step[];
    distance:         number;
    summary:          string;
}

export interface Admin {
    iso_3166_1_alpha3: string;
    iso_3166_1:        string;
}

export interface Annotation {
    duration: number[];
}

export interface Step {
    intersections:    Intersection[];
    maneuver:         Maneuver;
    name:             string;
    weight_typical:   number;
    duration_typical: number;
    duration:         number;
    distance:         number;
    driving_side:     string;
    weight:           number;
    mode:             string;
    ref?:             string;
    geometry:         Geometry;
}

export interface Intersection {
    entry:              boolean[];
    bearings:           number[];
    duration?:          number;
    stop_sign?:         boolean;
    mapbox_streets_v8?: MapboxStreetsV8;
    is_urban?:          boolean;
    admin_index:        number;
    out?:               number;
    weight?:            number;
    geometry_index:     number;
    location:           number[];
    in?:                number;
    turn_weight?:       number;
    turn_duration?:     number;
    traffic_signal?:    boolean;
    lanes?:             Lane[];
}

export interface Lane {
    indications:       string[];
    valid:             boolean;
    active:            boolean;
    valid_indication?: string;
}

export interface MapboxStreetsV8 {
    class: Class;
}

export enum Class {
    Primary = "primary",
    Tertiary = "tertiary",
    TertiaryLink = "tertiary_link",
}

export interface Maneuver {
    type:           string;
    instruction:    string;
    bearing_after:  number;
    bearing_before: number;
    location:       number[];
    modifier?:      string;
}

export interface Waypoint {
    distance: number;
    name:     string;
    location: number[];
}